+++ CHANGELOG.md
@@ -11,3 +11,4 @@ Major changes since the last busboy release (0.31):
 * Tests were converted to Mocha (#11, #12, #22, #23)
 * Using the native TextDecoder and the package 'text-decoding' for fallback if nodejs does not support the requested encoding (#50)
 * Add isPartAFile-option, to make the file-detection configurable (#53)
+* Empty Parts will not hang the process (#55)